+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:universal_organizer/widgets.dart';
+import 'package:universal_organizer/base_class.dart';
 
 import 'package:hive_flutter/hive_flutter.dart';
