+++ lib/top_bar.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:universal_organizer/color_theme.dart';
-
 import 'package:hive/hive.dart';
 
 class TopBar extends StatefulWidget {
@@ -97,35 +96,6 @@ class Button extends StatelessWidget {
           color: theme.background.primary,
         ),
       ),
-    );
-  }
-}
-
-class BaseClass extends StatelessWidget {
-  final ValueNotifier<Brightness> brightness;
-
-  const BaseClass({
-    super.key,
-    required this.brightness,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return ValueListenableBuilder<Brightness>(
-      valueListenable: brightness,
-      builder: (context, brightnessValue, child) {
-        return MaterialApp(
-          theme: ThemeData(
-            useMaterial3: useMaterial3,
-            brightness: brightnessValue,
-            colorScheme: ColorScheme.fromSeed(
-              seedColor: colorSelected,
-              brightness: brightnessValue,
-            ),
-          ),
-          home: TopBar(brightness: brightness),
-        );
-      },
     );
   }
 }
